Using set path fixes weird priv_dir issues.
I'm not 100% sure how, but it seems that using `set_path` instead of
just `add_path` changes how libraries are looked into by the Erlang VM
when it comes to priv dirs and lib dirs.

This commit forces the addition of apps' paths through this function
rather than just adding the paths, which appears to fix some subtle bugs
in umbrella apps with private tests requiring access to priv and lib
dirs for the app they're part of.