Remove call to deprecated compiler
The new compiler modules remove all need for these to even be called in
the first place; depending on the `compile task` automatically sorts out
the compilation that this plugin requires.

This was likely a leftover of prior execution modes where we didn't
expect users to actually depend on PropEr and the modules wouldn't
compile on their own.